# backend/public/docs/openapi.yaml
openapi: 3.0.3
info:
    title: Laravel
    description: ''
    version: 1.0.0
servers:
    -
        url: 'http://localhost'
paths:
...
    /api/jabatan:
        get:
            summary: ''
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses:
                200:
                    description: ''
                    content:
                        application/json:
                            schema:
                                type: object
                                example:
                                    data: [{ id: 2, namaJabatan: Hrd, deskripsiJabatan: Hrd, golonganGaji: '2', created_at: '2024-11-05T14:40:08.000000Z', updated_at: '2024-11-05T14:40:08.000000Z', statusEnable: null }, { id: 4, namaJabatan: Hrd, deskripsiJabatan: Hrd, golonganGaji: '2', created_at: '2024-11-10T07:05:07.000000Z', updated_at: '2024-11-10T07:05:07.000000Z', statusEnable: null }, { id: 1, namaJabatan: Direktur, deskripsiJabatan: Direktur, golonganGaji: '1', created_at: '2024-11-06T00:01:14.000000Z', updated_at: '2024-11-06T00:01:35.000000Z', statusEnable: null }, { id: 5, namaJabatan: Hrd, deskripsiJabatan: Hrd, golonganGaji: '2', created_at: '2024-11-10T07:53:28.000000Z', updated_at: '2024-11-10T07:53:28.000000Z', statusEnable: true }, { id: 18, namaJabatan: 'Project Manager', deskripsiJabatan: PM, golonganGaji: '2', created_at: '2024-11-10T09:20:38.000000Z', updated_at: '2024-11-10T09:20:38.000000Z', statusEnable: true }, { id: 19, namaJabatan: as, deskripsiJabatan: as, golonganGaji: '2', created_at: '2024-11-10T09:27:03.000000Z', updated_at: '2024-11-10T09:27:03.000000Z', statusEnable: true }, { id: 20, namaJabatan: asas, deskripsiJabatan: assa, golonganGaji: '2', created_at: '2024-11-10T09:27:26.000000Z', updated_at: '2024-11-10T09:27:26.000000Z', statusEnable: true }, { id: 21, namaJabatan: asda, deskripsiJabatan: asda, golonganGaji: '2', created_at: '2024-11-10T09:28:42.000000Z', updated_at: '2024-11-10T09:28:42.000000Z', statusEnable: true }, { id: 22, namaJabatan: Programer, deskripsiJabatan: simrs, golonganGaji: '2', created_at: '2024-11-10T09:39:50.000000Z', updated_at: '2024-11-10T09:39:50.000000Z', statusEnable: true }]
                                    message: success
                                properties:
                                    data: { type: array, example: [{ id: 2, namaJabatan: Hrd, deskripsiJabatan: Hrd, golonganGaji: '2', created_at: '2024-11-05T14:40:08.000000Z', updated_at: '2024-11-05T14:40:08.000000Z', statusEnable: null }, { id: 4, namaJabatan: Hrd, deskripsiJabatan: Hrd, golonganGaji: '2', created_at: '2024-11-10T07:05:07.000000Z', updated_at: '2024-11-10T07:05:07.000000Z', statusEnable: null }, { id: 1, namaJabatan: Direktur, deskripsiJabatan: Direktur, golonganGaji: '1', created_at: '2024-11-06T00:01:14.000000Z', updated_at: '2024-11-06T00:01:35.000000Z', statusEnable: null }, { id: 5, namaJabatan: Hrd, deskripsiJabatan: Hrd, golonganGaji: '2', created_at: '2024-11-10T07:53:28.000000Z', updated_at: '2024-11-10T07:53:28.000000Z', statusEnable: true }, { id: 18, namaJabatan: 'Project Manager', deskripsiJabatan: PM, golonganGaji: '2', created_at: '2024-11-10T09:20:38.000000Z', updated_at: '2024-11-10T09:20:38.000000Z', statusEnable: true }, { id: 19, namaJabatan: as, deskripsiJabatan: as, golonganGaji: '2', created_at: '2024-11-10T09:27:03.000000Z', updated_at: '2024-11-10T09:27:03.000000Z', statusEnable: true }, { id: 20, namaJabatan: asas, deskripsiJabatan: assa, golonganGaji: '2', created_at: '2024-11-10T09:27:26.000000Z', updated_at: '2024-11-10T09:27:26.000000Z', statusEnable: true }, { id: 21, namaJabatan: asda, deskripsiJabatan: asda, golonganGaji: '2', created_at: '2024-11-10T09:28:42.000000Z', updated_at: '2024-11-10T09:28:42.000000Z', statusEnable: true }, { id: 22, namaJabatan: Programer, deskripsiJabatan: simrs, golonganGaji: '2', created_at: '2024-11-10T09:39:50.000000Z', updated_at: '2024-11-10T09:39:50.000000Z', statusEnable: true }], items: { type: object } }
                                    message: { type: string, example: success }
            tags:
                - Endpoints
            security: []
        post:
            summary: ''
            description: ''
            parameters:
                -
                    in: header
                    name: Content-Type
                    description: ''
                    example: application/json
                    schema:
                        type: string
                -
                    in: header
                    name: Accept
                    description: ''
                    example: application/json
                    schema:
                        type: string
            responses: {  }
            tags:
                - Endpoints
            security: []
...